+++ virtual_finance_api/endpoints/yahoo/types.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from enum import Enum
+
+
+class AdjustType(str, Enum):
+    auto = "auto"
+    backadjust = "backadjust"